CREATE MODULAR FILE FOR FUNCTIONS
